+++ src/main/java/dev/amble/ait/core/blockentities/ArtronCollectorBlockEntity.java
@@ -1,5 +1,8 @@
 package dev.amble.ait.core.blockentities;
 
+import dev.amble.ait.AITMod;
+import dev.amble.lib.animation.BedrockModelProvider;
+import dev.amble.lib.client.bedrock.BedrockModelReference;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.entity.BlockEntityTicker;
@@ -23,8 +26,9 @@ import dev.amble.ait.core.AITItems;
 import dev.amble.ait.core.engine.link.block.FluidLinkBlockEntity;
 import dev.amble.ait.core.item.ArtronCollectorItem;
 import dev.amble.ait.core.world.RiftChunkManager;
+import org.jetbrains.annotations.Nullable;
 
-public class ArtronCollectorBlockEntity extends FluidLinkBlockEntity implements BlockEntityTicker<ArtronCollectorBlockEntity>, ArtronHolder {
+public class ArtronCollectorBlockEntity extends FluidLinkBlockEntity implements BedrockModelProvider, BlockEntityTicker<ArtronCollectorBlockEntity>, ArtronHolder {
 
     public static final int FLOW_AMOUNT = 3;
 
@@ -163,5 +167,15 @@ public class ArtronCollectorBlockEntity extends FluidLinkBlockEntity implements 
             return;
 
         this.world.updateListeners(this.getPos(), this.getCachedState(), this.getCachedState(), Block.NOTIFY_LISTENERS);
+    }
+
+    @Override
+    public String getTexturePrefix() {
+        return "blockentities/machines";
+    }
+
+    @Override
+    public @Nullable BedrockModelReference getModel() {
+        return new BedrockModelReference(AITMod.MOD_ID, "artron_collector");
     }
 }