+++ src/main/java/dev/amble/ait/core/blocks/ArtronCollectorBlock.java
@@ -4,13 +4,10 @@ import static dev.amble.ait.client.util.TooltipUtil.addShiftHiddenTooltip;
 
 import java.util.List;
 
+import net.minecraft.block.*;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockEntityProvider;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.ShapeContext;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
@@ -34,7 +31,7 @@ import dev.amble.ait.core.engine.link.block.HorizontalFluidLinkBlock;
 @SuppressWarnings("deprecation")
 public class ArtronCollectorBlock extends HorizontalFluidLinkBlock implements BlockEntityProvider {
 
-    protected static final VoxelShape Y_SHAPE = Block.createCuboidShape(6.0, 0.0, 6.0, 10.0, 29.0, 10.0);
+    protected static final VoxelShape Y_SHAPE = Block.createCuboidShape(0.0, 0.0, 0.0, 16.0, 32.0, 16.0);
 
     public ArtronCollectorBlock(Settings settings) {
         super(settings);
@@ -48,6 +45,16 @@ public class ArtronCollectorBlock extends HorizontalFluidLinkBlock implements Bl
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
         return Y_SHAPE;
+    }
+
+    @Override
+    public BlockRenderType getRenderType(BlockState state) {
+        return BlockRenderType.ENTITYBLOCK_ANIMATED;
+    }
+
+    @Override
+    public boolean hasSidedTransparency(BlockState state) {
+        return true;
     }
 
     @Override