+++ CHANGELOG.md
@@ -191,3 +191,4 @@
 - fix: sonic screwdriver wouldnt tell you the tool or material type you need to break a block | by [@Loqor](https://github.com/Loqor) ([#2181](https://github.com/amblelabs/ait/pull/2181))
 - fix: Updated Spanish translations | by [@Peter1169](https://github.com/Peter1169) ([#2124](https://github.com/amblelabs/ait/pull/2124))
 - feat: Sonic can shear sheep in interaction mode! | ([#2182](https://github.com/amblelabs/ait/pull/2182))
+- feat: new artron collector model! | by [@Loqor](https://github.com/Loqor) ([#2187](https://github.com/amblelabs/ait/pull/2187))